refactored main layout stuff
