+++ app/imports/ui/components/layout/Footer.jsx
