+++ app/imports/ui/components/layout/NavBar.jsx
