+++ app/imports/ui/layouts/App.jsx
@@ -4,8 +4,8 @@ import { Meteor } from 'meteor/meteor';
 import 'semantic-ui-css/semantic.css';
 import { Roles } from 'meteor/alanning:roles';
 import { HashRouter as Router, Route, Switch, Redirect } from 'react-router-dom';
-import NavBar from '../components/NavBar';
-import Footer from '../components/Footer';
+import NavBar from '../components/layout/NavBar';
+import Footer from '../components/layout/Footer';
 import Landing from '../pages/Landing';
 import ListStudySessionsAll from '../pages/ListStudySessionsAll';
 import Sessions from '../pages/Sessions';
